feat(api): Add configurable limit on paginated request.
This add the query parameter 'limit' on both GET /transactions and GET /accounts endpoints.
The parameter is capped to a value of 1000.
The default value still the same as before (15).

The PR also fix some tests around pagination which was working but was not doing the right thing (see api/controllers/pagination_test.go line 209).